+++ browser/modules/sqlQuery.js
@@ -41,6 +41,8 @@ let _self = false;
 
 let editingIsEnabled = false;
 
+let draggableEnabled = false;
+
 let template;
 
 let elementPrefix;
@@ -184,6 +186,10 @@ module.exports = {
             editingIsEnabled = true;
         }
 
+        if (window.vidiConfig.popupDraggable) {
+            draggableEnabled = true;
+        }
+
         qStoreShadow = qstore;
 
         this.reset(qstore);
@@ -256,10 +262,24 @@ module.exports = {
             if (featureInfoTableOnMap && !backArrowIsAdded) {
                 backArrowIsAdded = true;
                 defaultTemplate = `
-                                <div class='show-when-multiple-hits' style='cursor: pointer;' onclick='javascript:$("#modal-info-body").show();$("#alternative-info-container").hide();$("#click-for-info-slide .modal-title").empty();'>
+                                <div class='show-when-multiple-hits' style='cursor: pointer;'>
                                     <span class='material-icons'  style=''>keyboard_arrow_left </span>
                                     <span style="top: -7px;position: relative;">${__("Back")}</span>
                                 </div>` + defaultTemplate;
+                $(document).arrive('.show-when-multiple-hits', function (e, data) {
+                    $(this).on('click', function (e) {
+                        $("#modal-info-body").show();
+                        $("#alternative-info-container").hide();
+                        $("#click-for-info-slide .modal-title").empty();
+                        _self.getQstore()?.forEach(store => {
+                            $.each(store.layer._layers, function (i, v) {
+                                if (store.layer && store.layer.resetStyle) {
+                                    store.layer.resetStyle(v);
+                                }
+                            });
+                        })
+                    })
+                })
             }
 
             if (parsedMeta.info_element_selector) {
@@ -309,7 +329,7 @@ module.exports = {
 
                             // Add alternative-info-container to pop-up if featureInfoTableOnMap or else in left slide panel
                             if (featureInfoTableOnMap) {
-                                L.popup({
+                                var popup = L.popup({
                                     minWidth: 350
                                 })
                                     .setLatLng(infoClickPoint)
@@ -322,6 +342,11 @@ module.exports = {
                                             editingStarted = false;
                                         }
                                     });
+
+                                if (draggableEnabled) {
+                                    _self.makeDraggable(popup);
+                                }
+
                                 $("#info-box-pop-up").html(popUpInner);
 
                             } else {
@@ -354,11 +379,12 @@ module.exports = {
 
                         // Set select_function if featureInfoTableOnMap = true
                         if ((typeof parsedMeta.select_function === "undefined" || parsedMeta.select_function === "") && featureInfoTableOnMap) {
-                            let selectFunction = `function(id, layer, key, sqlQuery){
-                                                     $("#modal-info-body").hide();
-                                                     $("#alternative-info-container").show();
-                                                  }`;
-                            selectCallBack = Function('"use strict";return (' + selectFunction + ')')();
+
+                            selectCallBack = function (id, layer, key, sqlQuery) {
+
+                                $("#modal-info-body").hide();
+                                $("#alternative-info-container").show();
+                            };
                         } else if (typeof parsedMeta.select_function !== "undefined" && parsedMeta.select_function !== "") {
                             try {
                                 selectCallBack = Function('"use strict";return (' + parsedMeta.select_function + ')')();
@@ -407,6 +433,11 @@ module.exports = {
                                         }
                                     }
                                 }
+
+                                if (draggableEnabled) {
+                                    _self.makeDraggable(popup);
+                                }
+
                                 setTimeout(() => {
                                     if (editingIsEnabled && layerIsEditable) {
                                         $(".gc2-edit-tools").css(`display`, `inline`);
@@ -824,5 +855,33 @@ module.exports = {
             $("#click-for-info-slide .modal-title").html(title);
 
         }
+    },
+
+    /**
+     * makes popup draggable
+     * @param popup {object}
+     */
+    makeDraggable: (popup) => {
+        const map = cloud.get().map
+        const draggable = new L.Draggable(popup._container, popup._wrapper);
+        // change cursor class
+        $(".leaflet-popup-content-wrapper").css('cursor', 'move');
+        draggable.on('dragstart', function (e) {
+            //on first drag, remove the pop-up tip
+            $(".leaflet-popup-tip-container").hide();
+        });
+        draggable.on('dragend', function (e) {
+            // set the new position
+            popup.setLatLng(map.layerPointToLatLng(e.target._newPos));
+        });
+        draggable.enable();
+    },
+
+    /**
+     * Get query stores
+     * @returns array
+     */
+    getQstore: () => {
+        return qStoreShadow;
     }
 };