+++ public/js/lib/leaflet/leaflet-src.esm.js
@@ -10011,7 +10011,6 @@ var Popup = DivOverlay.extend({
 			var closeButton = this._closeButton = create$1('a', prefix + '-close-button', container);
 			closeButton.setAttribute('role', 'button'); // overrides the implicit role=link of <a> elements #7399
 			closeButton.setAttribute('aria-label', 'Close popup');
-			closeButton.href = '#close';
 			closeButton.innerHTML = '<span aria-hidden="true">&#215;</span>';
 
 			on(closeButton, 'click', this.close, this);